third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/font-stretch-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'font-stretch' to the 'normal' keyword: normal
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'font-stretch' to the 'ultra-condensed' keyword: ultra-condensed
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'font-stretch' to the 'extra-condensed' keyword: extra-condensed
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'font-stretch' to the 'condensed' keyword: condensed
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'font-stretch' to the 'semi-condensed' keyword: semi-condensed
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'font-stretch' to the 'semi-expanded' keyword: semi-expanded
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'font-stretch' to the 'expanded' keyword: expanded
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'font-stretch' to the 'extra-expanded' keyword: extra-expanded
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'font-stretch' to the 'ultra-expanded' keyword: ultra-expanded
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'font-stretch' to a percent: -3.14%
  assert_approx_equals: expected 0 +/- 0.000001 but got -3
[FAIL] Can set 'font-stretch' to a percent: 3.14%
  assert_approx_equals: expected 3.14 +/- 0.000001 but got 3
Harness: the test ran to completion.
